+++ src/constants/providerFactory.constants.ts
@@ -1,11 +1,15 @@
-import { ProviderTypeEnum } from 'providers/types/providerFactory.types';
+import {
+  ProviderTypeEnum,
+  ProviderType
+} from 'providers/types/providerFactory.types';
 
-export const providerLabels: Record<string, string> = {
+export const providerLabels: Record<ProviderType, string> = {
   [ProviderTypeEnum.crossWindow]: 'MultiversX Web Wallet',
   [ProviderTypeEnum.extension]: 'MultiversX Wallet Extension',
   [ProviderTypeEnum.walletConnect]: 'xPortal App',
   [ProviderTypeEnum.ledger]: 'Ledger',
   [ProviderTypeEnum.metamask]: 'MetaMask Snap',
   [ProviderTypeEnum.passkey]: 'Passkey',
+  [ProviderTypeEnum.webview]: 'Webview',
   [ProviderTypeEnum.none]: ''
 };