+++ src/managers/UnlockPanelManager/UnlockPanelManager.ts
@@ -5,12 +5,13 @@ import { ProviderFactory } from 'providers/ProviderFactory';
 import {
   IProviderBase,
   IProviderFactory,
-  ProviderTypeEnum
+  ICustomProvider,
+  ProviderTypeEnum,
+  ProviderType
 } from 'providers/types/providerFactory.types';
 import { networkSelector } from 'store/selectors';
 import { getState } from 'store/store';
 import {
-  AllowedProviderType,
   OnCloseUnlockPanelType,
   LoginHandlerType,
   UnlockPanelEventsEnum,
@@ -36,7 +37,7 @@ export class UnlockPanelManager extends SidePanelBaseManager<
   private static instance: UnlockPanelManager;
   private static loginHandler: LoginHandlerType | null = null;
   private static onClose: OnCloseUnlockPanelType | null = null;
-  private static allowedProviders?: AllowedProviderType[] | null = null;
+  private static allowedProviders?: ICustomProvider[] | null = null;
 
   constructor() {
     super({
@@ -69,7 +70,7 @@ export class UnlockPanelManager extends SidePanelBaseManager<
     const { walletAddress } = networkSelector(getState());
 
     this.data = {
-      providers: UnlockPanelManager.getProvidersList(),
+      providers: this.getProvidersList(),
       walletAddress
     };
 
@@ -136,26 +137,39 @@ export class UnlockPanelManager extends SidePanelBaseManager<
     return takesZeroArguments;
   };
 
-  private static getProvidersList(): IProviderBase[] {
+  private getProvidersList(): IProviderBase[] {
     const customProviders = ProviderFactory.customProviders;
 
-    const enumProviderTypes = Object.values(ProviderTypeEnum).filter(
+    const defaultProviderTypes = Object.values(ProviderTypeEnum).filter(
       (type) =>
         type !== ProviderTypeEnum.none && type !== ProviderTypeEnum.webview
     );
 
     const allAvailableProviderTypes = [
-      ...enumProviderTypes,
+      ...defaultProviderTypes,
       ...customProviders.map((p) => p.type)
     ];
 
-    const allowedProviderTypes = this.allowedProviders
-      ? this.allowedProviders.filter((type) =>
-          allAvailableProviderTypes.includes(type)
-        )
+    const customProviderLabels = customProviders.reduce(
+      (acc, provider) => {
+        acc[provider.type] = provider.name;
+        return acc;
+      },
+      {} as Record<ProviderType, string>
+    );
+
+    const allAvailableLabels = {
+      ...providerLabels,
+      ...customProviderLabels
+    };
+
+    const allowedProviderTypes = UnlockPanelManager.allowedProviders
+      ? UnlockPanelManager?.allowedProviders
+          .map((p) => p.type)
+          .filter((type) => allAvailableProviderTypes.includes(type))
       : allAvailableProviderTypes;
 
-    const providerList: IProviderBase[] = allowedProviderTypes.map((type) => {
+    const providerList = allowedProviderTypes.map((type) => {
       const custom = customProviders.find(
         (customProvider) => customProvider.type === type
       );
@@ -164,7 +178,7 @@ export class UnlockPanelManager extends SidePanelBaseManager<
       }
 
       return {
-        name: providerLabels[type as ProviderTypeEnum] ?? type,
+        name: type in allAvailableLabels ? allAvailableLabels[type] : type,
         type
       };
     });