+++ src/managers/UnlockPanelManager/UnlockPanelManager.types.ts
@@ -1,7 +1,7 @@
 import {
   ICustomProvider,
   IProviderFactory,
-  ProviderTypeEnum
+  ProviderType
 } from 'providers/types/providerFactory.types';
 
 export enum UnlockPanelEventsEnum {
@@ -21,18 +21,17 @@ export type LoginFunctonType = ({
   anchor
 }: IProviderFactory) => Promise<void>;
 
-export type AllowedProviderType = ICustomProvider['type'];
-
-export type CustomProviderViewType<
-  T extends ProviderTypeEnum = ProviderTypeEnum
-> = Omit<ICustomProvider<T>, 'constructor'>;
+export type CustomProviderViewType<T extends ProviderType = ProviderType> =
+  Omit<ICustomProvider<T>, 'constructor'>;
 
 export type LoginCallbackType = () => void;
 
 export type LoginHandlerType = LoginFunctonType | LoginCallbackType;
 export type OnCloseUnlockPanelType = () => void;
 
-export type UnlockPanelManagerInitParamsType = {
+export type UnlockPanelManagerInitParamsType<
+  T extends ProviderType = ProviderType
+> = {
   /**
      * Accepts:
      * - 1️⃣ a `callback` that is executed after login is performed
@@ -58,7 +57,7 @@ export type UnlockPanelManagerInitParamsType = {
   /**
    * List of allowed providers
    */
-  allowedProviders?: AllowedProviderType[] | null;
+  allowedProviders?: ICustomProvider<T>[] | null;
   /**
    * Callback function to handle UI behavior when the unlock panel is closed
    * without completing the login process.