+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- [Refactor `refreshNativeAuthTokenLogin` to return netivateAuthToken](https://github.com/multiversx/mx-sdk-dapp/pull/1450)
+- [Refactor `refreshNativeAuthTokenLogin` to return netivateAuthToken](https://github.com/multiversx/mx-sdk-dapp/pull/1451)
+- [Updated ProviderTypesEnum from enum to const](https://github.com/multiversx/mx-sdk-dapp/pull/1450)
 
 ## [[5.0.0-alpha.6](https://github.com/multiversx/mx-sdk-dapp/pull/1449)] - 2025-06-10